campus_safety_platform/DataGen/requirements.txt
Django==5.0
djangorestframework
channels==4.1.0        # Django Channels for WebSocket support
daphne==4.0.0         # ASGI server for Django Channels
channels-redis
psycopg2-binary
kafka-python
plotly
pandas
numpy
dj-database-url
kafka-python==2.0.2
psycopg2-binary==2.9.10
requests==2.31.0
python-dotenv==1.0.0